Disable `supports_norm_batch` for GPS, since `GPSConv` does not handle `batch_size`
Indeed, to be compatible with `BasicGNN` `GPSConv` should forward `batch_size` to norm layer. Instead, it fails to capture the argument which leads to `batch_size` being forwarded to MPNN layer, which does not expect it.
Given we made `batch` required to fix self-attention within graphs, `batch` is still available for norm layers, and `batch_size` is unneeded.