Restore pre-transform option for PERSEVERE dataset
The restored pre-transform is not saved to disk (since the dataset is still read directly from JSON files) but it is applied only once on load, rather than on every access like transforms.
It was restored to enable an option for costly but constant transforms to be computed only once.